+++ application.meta.ts
@@ -21,11 +21,11 @@
 
 const Application = {
   name: "pup",
-  version: "1.0.0-rc.37",
+  version: "1.0.0-rc.38",
   url: "jsr:@pup/pup@$VERSION",
   canary_url: "https://raw.githubusercontent.com/Hexagon/pup/main/pup.ts",
   deno: null, /* Minimum stable version of Deno required to run Pup (without --unstable-* flags)  */
-  deno_unstable: "1.42.0", /* Minimum version of Deno required to run Pup (with --unstable-* flags) */
+  deno_unstable: "1.43.0", /* Minimum version of Deno required to run Pup (with --unstable-* flags) */
   repository: "https://github.com/hexagon/pup",
   changelog: "https://hexagon.github.io/pup/changelog.html",
   permissions: [